Change longs to ints where possible in internal datatypes. Bug: 266809684 atest CtsHealthFitnessDeviceTestCases

+++ framework/java/android/health/connect/datatypes/RestingHeartRateRecord.java
@@ -181,7 +181,7 @@ public final class RestingHeartRateRecord extends InstantRecord {
                                 .setDeviceType(getMetadata().getDevice().getType());
         recordInternal.setTime(getTime().toEpochMilli());
         recordInternal.setZoneOffset(getZoneOffset().getTotalSeconds());
-        recordInternal.setBeatsPerMinute(mBeatsPerMinute);
+        recordInternal.setBeatsPerMinute((int) mBeatsPerMinute);
         return recordInternal;
     }
 }
